requirements.txt
python-telegram-bot==20.7
selenium==4.15.2
undetected-chromedriver==3.5.4
beautifulsoup4==4.12.2
requests==2.31.0
psutil==5.9.6
python-dotenv==1.0.0
